refactor: 컨트롤러 테스트 책임 분리
